Switch third sentence in hero from p element to h3
I had made it a p element and put an id tag on it like the second sentence, but I wanted to make them all headings. I think that makes the most sense, and if someone is using a screen reader, all three sentences will get picked up. So I switched it to an h3 and then styled it like body copy.