Place images and text in divs and add "image" or "text-content" class
I wrapped each image that starts a new section in a div and added the "image" class. I also wrapped each block of text content (including blocks that have an image sandwiched between paragraphs, as longs as it's still in the same section) in a div and added the "text-content" class.